--- a/src/Grpc.Net.Client/GrpcChannel.cs
+++ b/src/Grpc.Net.Client/GrpcChannel.cs
@@ -40,9 +40,11 @@ namespace Grpc.Net.Client
 
         private readonly ConcurrentDictionary<IMethod, GrpcMethodInfo> _methodInfoCache;
         private readonly Func<IMethod, GrpcMethodInfo> _createMethodInfoFunc;
+        // Internal for testing
+        internal readonly HashSet<IDisposable> ActiveCalls;
 
         internal Uri Address { get; }
-        internal HttpClient HttpClient { get; }
+        internal HttpMessageInvoker HttpInvoker { get; }
         internal int? SendMaxMessageSize { get; }
         internal int? ReceiveMaxMessageSize { get; }
         internal ILoggerFactory LoggerFactory { get; }
@@ -64,12 +66,13 @@ namespace Grpc.Net.Client
             _methodInfoCache = new ConcurrentDictionary<IMethod, GrpcMethodInfo>();
 
             // Dispose the HttpClient if...
-            //   1. No client was specified and so the channel created the HttpClient itself
-            //   2. User has specified a client and set DisposeHttpClient to true
-            _shouldDisposeHttpClient = channelOptions.HttpClient == null || channelOptions.DisposeHttpClient;
+            //   1. No client/handler was specified and so the channel created the client itself
+            //   2. User has specified a client/handler and set DisposeHttpClient to true
+            _shouldDisposeHttpClient = (channelOptions.HttpClient == null && channelOptions.HttpHandler == null)
+                || channelOptions.DisposeHttpClient;
 
             Address = address;
-            HttpClient = channelOptions.HttpClient ?? CreateInternalHttpClient();
+            HttpInvoker = channelOptions.HttpClient ?? CreateInternalHttpInvoker(channelOptions.HttpHandler);
             SendMaxMessageSize = channelOptions.MaxSendMessageSize;
             ReceiveMaxMessageSize = channelOptions.MaxReceiveMessageSize;
             CompressionProviders = ResolveCompressionProviders(channelOptions.CompressionProviders);
@@ -77,6 +80,7 @@ namespace Grpc.Net.Client
             LoggerFactory = channelOptions.LoggerFactory ?? NullLoggerFactory.Instance;
             ThrowOperationCanceledOnCancellation = channelOptions.ThrowOperationCanceledOnCancellation;
             _createMethodInfoFunc = CreateMethodInfo;
+            ActiveCalls = new HashSet<IDisposable>();
 
             if (channelOptions.Credentials != null)
             {
@@ -90,21 +94,29 @@ namespace Grpc.Net.Client
             }
         }
 
-        private static HttpClient CreateInternalHttpClient()
+        private static HttpMessageInvoker CreateInternalHttpInvoker(HttpMessageHandler? handler)
         {
-            var httpClient = new HttpClient();
+            // HttpMessageInvoker should always dispose handler if Disposed is called on it.
+            // Decision to dispose invoker is controlled by _shouldDisposeHttpClient.
+            var httpInvoker = new HttpMessageInvoker(handler ?? new HttpClientHandler(), disposeHandler: true);
 
-            // Long running server and duplex streaming gRPC requests may not
-            // return any messages for over 100 seconds, triggering a cancellation
-            // of HttpClient.SendAsync. Disable timeout in internally created
-            // HttpClient for channel.
-            //
-            // gRPC deadline should be the recommended way to timeout gRPC calls.
-            //
-            // https://github.com/dotnet/corefx/issues/41650
-            httpClient.Timeout = Timeout.InfiniteTimeSpan;
+            return httpInvoker;
+        }
 
-            return httpClient;
+        internal void RegisterActiveCall(IDisposable grpcCall)
+        {
+            lock (ActiveCalls)
+            {
+                ActiveCalls.Add(grpcCall);
+            }
+        }
+
+        internal void FinishActiveCall(IDisposable grpcCall)
+        {
+            lock (ActiveCalls)
+            {
+                ActiveCalls.Remove(grpcCall);
+            }
         }
 
         internal GrpcMethodInfo GetCachedGrpcMethodInfo(IMethod method)
@@ -261,6 +273,12 @@ namespace Grpc.Net.Client
                 throw new ArgumentNullException(nameof(channelOptions));
             }
 
+            if (channelOptions.HttpClient != null && channelOptions.HttpHandler != null)
+            {
+                throw new ArgumentException($"{nameof(GrpcChannelOptions.HttpClient)} and {nameof(GrpcChannelOptions.HttpHandler)} have been configured. " +
+                    $"Only one HTTP caller can be specified.");
+            }
+
             return new GrpcChannel(address, channelOptions);
         }
 
@@ -275,9 +293,18 @@ namespace Grpc.Net.Client
                 return;
             }
 
+            lock (ActiveCalls)
+            {
+                // Disposing calls will remove them from ActiveCalls
+                foreach (var activeCall in ActiveCalls)
+                {
+                    activeCall.Dispose();
+                }
+            }
+
             if (_shouldDisposeHttpClient)
             {
-                HttpClient.Dispose();
+                HttpInvoker.Dispose();
             }
             Disposed = true;
         }