--- a/src/Grpc.Net.Client/Internal/GrpcCall.cs
+++ b/src/Grpc.Net.Client/Internal/GrpcCall.cs
@@ -170,6 +170,8 @@ namespace Grpc.Net.Client.Internal
                 ClientStreamReader?.HttpResponseTcs.TrySetCanceled();
             }
 
+            Channel.FinishActiveCall(this);
+
             _ctsRegistration?.Dispose();
             _deadlineTimer?.Dispose();
             HttpResponse?.Dispose();