--- a/test/FunctionalTests/Client/CancellationTests.cs
+++ b/test/FunctionalTests/Client/CancellationTests.cs
@@ -191,6 +191,75 @@ namespace Grpc.AspNetCore.FunctionalTests.Client
         }
 
         [Test]
+        public async Task ServerStreaming_ChannelDisposed_CancellationSentToServer()
+        {
+            var syncPoint = new SyncPoint();
+            var serverCompleteTcs = new TaskCompletionSource<object?>(TaskCreationOptions.RunContinuationsAsynchronously);
+
+            async Task ServerStreamingCall(DataMessage request, IServerStreamWriter<DataMessage> streamWriter, ServerCallContext context)
+            {
+                await syncPoint.WaitToContinue().DefaultTimeout();
+
+                // Wait until the client cancels
+                while (!context.CancellationToken.IsCancellationRequested)
+                {
+                    await Task.Delay(TimeSpan.FromMilliseconds(10));
+                }
+
+                serverCompleteTcs.TrySetResult(null);
+            }
+
+            // Arrange
+            SetExpectedErrorsFilter(writeContext =>
+            {
+                // Kestrel cancellation error message
+                if (writeContext.Exception is IOException &&
+                    writeContext.Exception.Message == "The client reset the request stream.")
+                {
+                    return true;
+                }
+
+                if (writeContext.LoggerName == "Grpc.Net.Client.Internal.GrpcCall" &&
+                    writeContext.EventId.Name == "ErrorStartingCall" &&
+                    writeContext.Message == "Error starting gRPC call.")
+                {
+                    return true;
+                }
+
+                // Ignore all logging related errors for now
+                return false;
+            });
+
+            var method = Fixture.DynamicGrpc.AddServerStreamingMethod<DataMessage, DataMessage>(ServerStreamingCall);
+
+            var channel = CreateChannel(useHandler: true);
+
+            var client = TestClientFactory.Create(channel, method);
+
+            // Act
+            var call = client.ServerStreamingCall(new DataMessage());
+            await syncPoint.WaitForSyncPoint();
+            syncPoint.Continue();
+
+            // Assert
+            Assert.AreEqual(1, channel.ActiveCalls.Count);
+            var moveNextTask = call.ResponseStream.MoveNext(CancellationToken.None);
+
+            channel.Dispose();
+
+            Assert.AreEqual(0, channel.ActiveCalls.Count);
+
+            var ex = await ExceptionAssert.ThrowsAsync<RpcException>(() => moveNextTask).DefaultTimeout();
+            Assert.AreEqual(StatusCode.Cancelled, ex.StatusCode);
+
+            await serverCompleteTcs.Task.DefaultTimeout();
+
+            await TestHelpers.AssertIsTrueRetryAsync(
+                () => HasLog(LogLevel.Information, "GrpcStatusError", "Call failed with gRPC error status. Status code: 'Cancelled', Message: 'gRPC call disposed.'."),
+                "Missing client cancellation log.").DefaultTimeout();
+        }
+
+        [Test]
         public async Task ServerStreaming_CancellationOnClientAfterResponseHeadersReceived_CancellationSentToServer()
         {
             var serverCompleteTcs = new TaskCompletionSource<object?>(TaskCreationOptions.RunContinuationsAsynchronously);