@stephentoub 2020-05-05
> 9.730 us => 7.860 us

Nice!